Find the First Extreme Peak or Trough Value in a Numeric Vector

Determines the direction of the response of y (positive or negative) and finds the first extreme value (peak for positive, or trough for negative) in which no greater extreme is observed within a trailing window in units of x defined by end_kinetics_span, or the limits of the data.

Usage

find_first_extreme(y, x = NULL, end_kinetics_span = 30)

Arguments

y

A numeric vector of the response variable, or the name of the variable within a dataframe.

x

An optional numeric vector of the predictor variable, or the name of the variable within a dataframe. Defaults to using the index of x = seq_along(y).

end_kinetics_span

(Default end_kinetics_span = 30) A numeric scalar indicating the local window in units of the predictor variable x after the kinetics extreme (peak or trough) value to look for subsequent greater extremes. The kinetics model will be fit to the data up to the first local extreme with no subsequent greater extremes within a trailing window in units of x defined by end_kinetics_span, or the limits of the data.

Value

A list with components:

x

The filtered x values from the origin to the end of the trailing window defined by end_kinetics_span past the first extreme value.

y

The filtered y values from the origin to the end of the trailing window defined by end_kinetics_span past the first extreme value.

extreme

A tibble with the x and y coordinates of the first extreme value, as columns.

See also