Process Kinetics

Fit mNIRS kinetics vector data with a parametric or non-parametric curve fitting model.

Usage

process_kinetics(
  x,
  y = NULL,
  data = NULL,
  x0 = 0,
  method = c("monoexponential", "sigmoidal", "half_time", "peak_slope"),
  ...
)

Arguments

x

A numeric vector specifying the predictor variable for y if y is defined. Otherwise, x is assumed to define the response variable.

y

(Optional). A numeric vector specifying the response variable. If y is not defined, the response variable is assumed to be defined by x, with idx = seq_along(x) as the predictor variable.

data

(Optional). A dataframe containing at least the response variable (x), or the predictor and response variables (x and y).

x0

A numeric scalar indicating the value of the predictor variable x or idx representing the start of the kinetics event (default x0 = 0).

method

Indicates how to process the kinetics.

method = "monoexponential"

...

method = "sigmoidal"

...

method = "half_time"

...

method = "peak_slope"

...

...

Additional arguments. Used to define fixed parameters which will not be optimised by the kinetics methods. e.g. A = 10 will define SSmonoexp(x, A = 10, B, TD, tau)

Value

A list L of class mNIRS.kinetics with components L$...:

model

The model object.

data

A dataframe of input and fitted model data.

fitted

A vector of fitted values returned by the model.

coefs

A vector of model coefficients, including manually fixed parameters.

fit_criteria

A vector of model fit criteria (AIC, BIC, R2, RMSE, RSE, MAE, MAPE).

Details

method %in% c("monoexponential", "sigmoidal") use nls() for nonlinear (weighted) least-squares estimates.

See also