Read mNIRS Data From File

Create an mNIRS dataframe imported from file.

Usage

read_data(
  file_path,
  nirs_columns,
  sample_column = NULL,
  event_column = NULL,
  sample_rate = NULL,
  numeric_time = TRUE,
  keep_all = FALSE,
  verbose = TRUE
)

Arguments

file_path

The file path including extension (either ".xlsx", ".xls", or ".csv") to import.

nirs_columns

A character vector indicating the mNIRS data columns to import from the target file. Must match exactly. A named character vector can be used to rename columns (see Details).

sample_column

(Optional). A character scalar indicating the name of a time or sample data column. Must match exactly. A named character vector can be used to rename columns.

event_column

(Optional). A character scalar indicating the name of an event or lap data column. Must match exactly. A named character vector can be used to rename columns.

sample_rate

(Optional). A numeric scalar for the sample rate in Hz. If not defined explicitly, will be estimated from the file data (see Details).

numeric_time

A logical. TRUE (default) will convert date-time formatted columns to numeric values in seconds. FALSE will retain these columns formatted as date-time in the format of the original file.

keep_all

A logical. FALSE (default) will only include the explicitly indicated data columns. TRUE will include all columns detected from the file.

verbose

A logical. TRUE (default) will return warnings and messages which can be used for data error checking. FALSE will silence these messages. Errors will always be returned.

Value

A tibble of class mNIRS.data with metadata available with attributes().

Details

Column names are matched to a single row, anywhere in the data file, not necessarily the top row of the file. Column names must match exactly to data file text. If there are duplicate column names, the columns will be matched in order of appearance and appended with ...X, where X is the column number. You may want to double check the correct columns have been assigned as intended.

Columns can be renamed in the format c("new_name" = "original_name"), where "original_name" should match the file text exactly.

sample_rate is required for certain {mNIRS} functions. If it is not defined explicitly, sample_rate will be estimated based on the mean difference between values in the sample_column. If sample_column is not defined, then sample_rate will be set to 1 Hz. If sample_column in the data file contains integer row numbers, then sample_rate will be incorrectly estimated to be 1 Hz, and should be defined explicitly.

Columns and rows which are entirely missing (NA) are omitted.