Read mNIRS Data From File

Will read files exported from most commercially available mNIRS devices and return a dataframe of class "mnirs.data" with recorded data and metadata.

Usage

read_data(
  file_path,
  nirs_channels,
  sample_channel = NULL,
  event_channel = NULL,
  sample_rate = NULL,
  time_to_numeric = TRUE,
  time_from_zero = FALSE,
  keep_all = FALSE,
  verbose = TRUE
)

Arguments

file_path

The file path including extension (either ".xlsx", ".xls", or ".csv") to import.

nirs_channels

A character vector indicating the mNIRS data columns to import from the file. Must match column names in the data file exactly. A named character vector can be used to rename columns in the form: c(new_name = "old_name") (see Details).

sample_channel

An optional character scalar indicating the time or sample data column to import from the file. Must match column names in the data file exactly. A named character vector can be used to rename columns in the form: c(new_name = "old_name") (see Details).

event_channel

An optional character scalar indicating the event or lap data column to import from the file. Must match column names in the data file exactly. A named character vector can be used to rename columns in the form: c(new_name = "original_name") (see Details).

sample_rate

An optional numeric scalar for the sample rate in Hz. If not defined explicitly, will be estimated from the data (see Details).

time_to_numeric

A logical. TRUE (the default) will convert a date-time formatted sample_channel to numeric values in seconds. FALSE will return sample_channel in the format of the original file.

time_from_zero

A logical. FALSE (the default) will return the original numeric values of sample_channel. TRUE will re-sample sample_channel to start from zero.

keep_all

A logical. FALSE (the default) will only include the explicitly specified data columns. TRUE will include all columns detected from the file.

verbose

A logical. TRUE (the default) will return warnings and messages which can be used for troubleshooting. FALSE will silence these messages. Errors will always be returned.

Value

A tibble of class "mnirs.data" with metadata available with attributes().

Details

Channel names are matched to a single row, representing the header row for data columns anywhere in the data file, not necessarily the top row of the file.

Channels can be renamed in the format c(new_name = "original_name"), where "original_name" should match the column names found in the file exactly.

If there are duplicate column names in the file, the channel names will be matched in the order in which they appear. You may want to confirm that the correct columns have been assigned to each channel as intended.

If sample_channel is not specified, then an index column will be added from the row numbers. If the specified sample_channel contains unequal sampling (i.e., repeated values or unordered samples) a warning will be given suggesting the user confirm the file data manually.

When the sample_channel is provided in date-time format (e.g. hh:mm:ss), this can be converted back to numeric values by time_to_numeric = TRUE. In this case, values will be recalculated as starting from 0 at the first sample.

sample_channel will typically represent time values in seconds. However, some NIRS devices export the sample index. This can be converted to time values if the sample_rate is known.

sample_rate is required for certain {mnirs} functions to work properly. If it is not defined explicitly, it will be estimated based on the mean difference between values in the sample_channel. If sample_channel is not defined, then sample_rate will be set to 1 Hz. If sample_channel in the data file contains integer row numbers, then sample_rate will be incorrectly estimated to be 1 Hz, and should be defined explicitly.

Columns and rows which contain entirely missing data (NA) are omitted.